backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
motor==3.7.0
pymongo==4.12.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt<4
email-validator==2.2.0
python-multipart==0.0.20
slowapi==0.1.9
cachetools==5.5.2
pandas==2.2.3
openpyxl==3.1.5
numpy==2.2.4
faiss-cpu==1.10.0
openai==1.75.0
httpx==0.28.1
orjson==3.10.16
groq==0.18.0
